# src/common/config/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "MIRA Football - Backend API",
    "description": "API hệ thống đặt sân bóng đá MIRA Football - By Trịnh Quốc Hùng",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/api",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "BearerAuth": {
      "type": "http",
      "scheme": "bearer",
      "bearerFormat": "JWT"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/upload/avatar": {
      "post": {
        "tags": [
          "Upload"
        ],
        "summary": "Upload ảnh đại diện",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/upload/single": {
      "post": {
        "tags": [
          "Upload"
        ],
        "summary": "Upload một file",
        "description": "",
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
  }
}